Se agrega observer de NC al AppServiceProvider & Migraciones faltantes

+++ database/migrations/2023_07_29_201033_remove_stock_from_articulos_table.php
@@ -14,7 +14,9 @@ class RemoveStockFromArticulosTable extends Migration
     public function up()
     {
         Schema::table('articulos', function (Blueprint $table) {
-            $table->dropColumn('stock');
+            if (Schema::hasColumn('articulos', 'stock')) {
+                $table->dropColumn('stock');
+            }
         });
     }
